backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
django-cors-headers==4.9.0

